requirements.txt
torch==2.4.1
torchvision==0.19.1
transformers==4.46.0
pillow==10.2.0
accelerate==0.34.0
peft==0.13.0
qwen-vl-utils
tqdm==4.66.4
